Working version of dynamodb persisting report content as json tree
