+++ pstatus-report-sink-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/ReportManager.kt
@@ -4,7 +4,6 @@ import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.ToNumberPolicy
 import com.google.gson.reflect.TypeToken
-import gov.cdc.ocio.processingstatusapi.dynamo.Content
 import gov.cdc.ocio.processingstatusapi.exceptions.BadRequestException
 import gov.cdc.ocio.processingstatusapi.exceptions.BadStateException
 import gov.cdc.ocio.processingstatusapi.models.DispositionType
@@ -17,6 +16,7 @@ import mu.KotlinLogging
 import org.apache.commons.lang3.SerializationUtils
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
+import software.amazon.awssdk.protocols.jsoncore.JsonNode
 import java.time.Instant
 import java.util.*
 
@@ -244,16 +244,16 @@ class ReportManager: KoinComponent {
             this.contentType = contentType
 
             if (contentType.lowercase() == "application/json" || contentType.lowercase() == "json") {
-                val typeObject = object : TypeToken<HashMap<*, *>?>() {}.type
-                val jsonMap: Map<String, Any> = gson.fromJson(Gson().toJson(content, MutableMap::class.java).toString(), typeObject)
-                this.content = Content()
-                jsonMap.entries.forEach {
-                    this.content!![it.key] = it.value
+                try {
+                    val json = Gson().toJson(content, MutableMap::class.java).toString()
+                    this.content = JsonNode.parser().parse(json)
+                } catch (e: Exception) {
+                    logger.error { e.localizedMessage }
                 }
             } else {
                 val bytesArray = SerializationUtils.serialize(content as String)
                 val base64Encoded = Base64.getEncoder().encodeToString(bytesArray)
-                this.content = mapOf("base64Encoded" to base64Encoded) as Content
+//                this.content = mapOf("base64Encoded" to base64Encoded) as Content
             }
         }
        return createReportItem(uploadId,stageReportId,stageReport)
@@ -320,8 +320,8 @@ class ReportManager: KoinComponent {
             this.validationSchemas= validationSchemaFileNames
             if (contentType?.lowercase() == "json" && !isNullOrEmpty(content) && !isBase64Encoded(content.toString())) {
                 val typeObject = object : TypeToken<HashMap<*, *>?>() {}.type
-                val jsonMap: Content = gson.fromJson(Gson().toJson(content, MutableMap::class.java).toString(), typeObject)
-                this.content = jsonMap
+//                val jsonMap: Content = gson.fromJson(Gson().toJson(content, MutableMap::class.java).toString(), typeObject)
+//                this.content = jsonMap
             } //else
                 //this.content = content
         }
